Pull all p2p unit tests out of replication tests and put them into p2p tests.
