+++ src/Couchbase.Lite.Tests.Shared/ReplicationTest.cs
@@ -59,17 +59,13 @@ namespace Test
 #endif
     public sealed class ReplicatorTest : TestCase
     {
-        const ushort WSPort = 0;
-        const ushort WSSPort = 0;
-
         private static int Counter;
         private Database _otherDB;
         private Replicator _repl;
         private WaitAssert _waitAssert;
         private bool _isFilteredCallback;
         private List<DocumentReplicationEventArgs> _replicationEvents = new List<DocumentReplicationEventArgs>();
-        URLEndpointListener _listener;
-        URLEndpointListenerConfiguration _config;
+
 #if !WINDOWS_UWP
         public ReplicatorTest(ITestOutputHelper output) : base(output)
 #else
@@ -105,7 +101,7 @@ namespace Test
             }
         }
 #endif
-#if COUCHBASE_ENTERPRISE
+
         [Fact]
         public void TestReadOnlyConfiguration()
         {
@@ -820,295 +816,6 @@ namespace Test
             callbackCount.Should().Be(0);
             _repl.Status.Progress.Total.Should().Be(0UL);
         }
-
-        #region p2p unit tests
-        
-        [Fact]
-        public void TestShortP2P()
-        {
-            //var testNo = 1;
-            foreach (var protocolType in new[] { ProtocolType.ByteStream, ProtocolType.MessageStream }) {
-                using (var mdoc = new MutableDocument("livesindb")) {
-                    mdoc.SetString("name", "db");
-                    Db.Save(mdoc);
-                }
-
-                using (var mdoc = new MutableDocument("livesinotherdb")) {
-                    mdoc.SetString("name", "otherdb");
-                    _otherDB.Save(mdoc);
-                }
-
-                var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, protocolType));
-                var server = new MockServerConnection(listener, protocolType);
-                var messageendpoint = new MessageEndpoint($"p2ptest1", server, protocolType,
-                        new MockConnectionFactory(null));
-                var uid = messageendpoint.Uid;
-
-                // PUSH
-                var config = new ReplicatorConfiguration(Db, messageendpoint) {
-                    ReplicatorType = ReplicatorType.Push,
-                    Continuous = false
-                };
-                RunReplication(config, 0, 0);
-                _otherDB.Count.Should().Be(2UL, "because it contains the original and new");
-                Db.Count.Should().Be(1UL, "because there is no pull, so the first db should only have the original");
-
-                // PULL
-                config = new ReplicatorConfiguration(Db, messageendpoint) {
-                    ReplicatorType = ReplicatorType.Pull,
-                    Continuous = false
-                };
-
-                RunReplication(config, 0, 0);
-                Db.Count.Should().Be(2UL, "because the pull should add the document from otherDB");
-
-                using (var savedDoc = Db.GetDocument("livesinotherdb"))
-                using (var mdoc = savedDoc.ToMutable()) {
-                    mdoc.SetBoolean("modified", true);
-                    Db.Save(mdoc);
-                }
-
-                using (var savedDoc = _otherDB.GetDocument("livesindb"))
-                using (var mdoc = savedDoc.ToMutable()) {
-                    mdoc.SetBoolean("modified", true);
-                    _otherDB.Save(mdoc);
-                }
-
-                // PUSH & PULL
-                config = new ReplicatorConfiguration(Db,
-                        new MessageEndpoint($"p2ptest1", server, protocolType,
-                            new MockConnectionFactory(null))) { Continuous = false };
-
-                RunReplication(config, 0, 0);
-                Db.Count.Should().Be(2UL, "because no new documents were added");
-
-                using (var savedDoc = Db.GetDocument("livesindb")) {
-                    savedDoc.GetBoolean("modified").Should()
-                        .BeTrue("because the property change should have come from the other DB");
-                }
-
-                using (var savedDoc = _otherDB.GetDocument("livesinotherdb")) {
-                    savedDoc.GetBoolean("modified").Should()
-                        .BeTrue("because the property change should come from the original DB");
-                }
-
-                Db.Delete();
-                ReopenDB();
-                _otherDB.Delete();
-                _otherDB.Dispose();
-                _otherDB = OpenDB(_otherDB.Name);
-            }
-        }
-
-        [Fact]
-        public void TestContinuousP2P()
-        {
-            _otherDB.Delete();
-            _otherDB = OpenDB(_otherDB.Name);
-            Db.Delete();
-            ReopenDB();
-            RunTwoStepContinuous(ReplicatorType.Push, "p2ptest1");
-            _otherDB.Delete();
-            _otherDB = OpenDB(_otherDB.Name);
-            Db.Delete();
-            ReopenDB();
-            RunTwoStepContinuous(ReplicatorType.Pull, "p2ptest2");
-            _otherDB.Delete();
-            _otherDB = OpenDB(_otherDB.Name);
-            Db.Delete();
-            ReopenDB();
-            RunTwoStepContinuous(ReplicatorType.PushAndPull, "p2ptest3");
-        }
-
-        //[Fact]
-        //public void TestP2PRecoverableFailureDuringOpen() => TestP2PError(MockConnectionLifecycleLocation.Connect, true);
-
-        //[Fact]
-        //public void TestP2PRecoverableFailureDuringSend() => TestP2PError(MockConnectionLifecycleLocation.Send, true);
-
-        //[Fact]
-        //public void TestP2PRecoverableFailureDuringReceive() => TestP2PError(MockConnectionLifecycleLocation.Receive, true);
-
-        [Fact]
-        public void TestP2PPermanentFailureDuringOpen() => TestP2PError(MockConnectionLifecycleLocation.Connect, false);
-
-        [Fact]
-        public void TestP2PPermanentFailureDuringSend() => TestP2PError(MockConnectionLifecycleLocation.Send, false);
-
-        [Fact]
-        public void TestP2PPermanentFailureDuringReceive() => TestP2PError(MockConnectionLifecycleLocation.Receive, false);
-
-        [Fact]
-        public void TestP2PFailureDuringClose()
-        {
-            using (var mdoc = new MutableDocument("livesindb")) {
-                mdoc.SetString("name", "db");
-                Db.Save(mdoc);
-            }
-
-            var config = CreateFailureP2PConfiguration(ProtocolType.ByteStream, MockConnectionLifecycleLocation.Close,
-                false);
-            RunReplication(config, (int)CouchbaseLiteError.WebSocketUserPermanent, CouchbaseLiteErrorType.CouchbaseLite);
-            config = CreateFailureP2PConfiguration(ProtocolType.MessageStream, MockConnectionLifecycleLocation.Close,
-                false);
-            RunReplication(config, (int)CouchbaseLiteError.WebSocketUserPermanent, CouchbaseLiteErrorType.CouchbaseLite, true);
-        }
-
-        [Fact]
-        public void TestP2PPassiveClose()
-        {
-            var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, ProtocolType.MessageStream));
-            var awaiter = new ListenerAwaiter(listener);
-            var serverConnection = new MockServerConnection(listener, ProtocolType.MessageStream);
-            var errorLogic = new ReconnectErrorLogic();
-            var config = new ReplicatorConfiguration(Db,
-                new MessageEndpoint("p2ptest1", serverConnection, ProtocolType.MessageStream,
-                    new MockConnectionFactory(errorLogic))) {
-                Continuous = true
-            };
-
-            using (var replicator = new Replicator(config)) {
-                replicator.Start();
-
-                var count = 0;
-                while (count++ < 10 && replicator.Status.Activity != ReplicatorActivityLevel.Idle) {
-                    Thread.Sleep(500);
-                    count.Should().BeLessThan(10, "because otherwise the replicator never went idle");
-                }
-                var connection = listener.Connections;
-                errorLogic.ErrorActive = true;
-                listener.Close(serverConnection);
-                count = 0;
-                while (count++ < 10 && replicator.Status.Activity != ReplicatorActivityLevel.Stopped) {
-                    Thread.Sleep(500);
-                    count.Should().BeLessThan(10, "because otherwise the replicator never stopped");
-                }
-
-                awaiter.WaitHandle.WaitOne(TimeSpan.FromSeconds(10)).Should().BeTrue();
-                awaiter.Validate();
-
-                replicator.Status.Error.Should()
-                    .NotBeNull("because closing the passive side creates an error on the active one");
-            }
-        }
-
-        [Fact]
-        public void TestP2PPassiveCloseAll()
-        {
-            using (var doc = new MutableDocument("test")) {
-                doc.SetString("name", "Smokey");
-                Db.Save(doc);
-            }
-
-            var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, ProtocolType.MessageStream));
-            var serverConnection1 = new MockServerConnection(listener, ProtocolType.MessageStream);
-            var serverConnection2 = new MockServerConnection(listener, ProtocolType.MessageStream);
-            var closeWait1 = new ManualResetEventSlim();
-            var closeWait2 = new ManualResetEventSlim();
-            var errorLogic = new ReconnectErrorLogic();
-            var config = new ReplicatorConfiguration(Db,
-                new MessageEndpoint("p2ptest1", serverConnection1, ProtocolType.MessageStream,
-                    new MockConnectionFactory(errorLogic))) {
-                Continuous = true
-            };
-
-            var config2 = new ReplicatorConfiguration(Db,
-                new MessageEndpoint("p2ptest2", serverConnection2, ProtocolType.MessageStream,
-                    new MockConnectionFactory(errorLogic))) {
-                Continuous = true
-            };
-
-            using (var replicator = new Replicator(config))
-            using (var replicator2 = new Replicator(config2)) {
-                replicator.Start();
-                replicator2.Start();
-
-                var count = 0;
-                while (count++ < 10 && replicator.Status.Activity != ReplicatorActivityLevel.Idle &&
-                       replicator2.Status.Activity != ReplicatorActivityLevel.Idle) {
-                    Thread.Sleep(500);
-                    count.Should().BeLessThan(10, "because otherwise the replicator(s) never went idle");
-                }
-
-                errorLogic.ErrorActive = true;
-                listener.AddChangeListener((sender, args) => {
-                    if (args.Status.Activity == ReplicatorActivityLevel.Stopped) {
-                        if (args.Connection == serverConnection1) {
-                            closeWait1.Set();
-                        } else {
-                            closeWait2.Set();
-                        }
-                    }
-                });
-                var connection = listener.Connections;
-                listener.CloseAll();
-                count = 0;
-                while (count++ < 10 && replicator.Status.Activity != ReplicatorActivityLevel.Stopped &&
-                       replicator2.Status.Activity != ReplicatorActivityLevel.Stopped) {
-                    Thread.Sleep(500);
-                    count.Should().BeLessThan(10, "because otherwise the replicator(s) never stopped");
-                }
-
-                closeWait1.Wait(TimeSpan.FromSeconds(5)).Should()
-                    .BeTrue("because otherwise the first listener did not stop");
-                closeWait2.Wait(TimeSpan.FromSeconds(5)).Should()
-                    .BeTrue("because otherwise the second listener did not stop");
-
-                replicator.Status.Error.Should()
-                    .NotBeNull("because closing the passive side creates an error on the active one");
-                replicator2.Status.Error.Should()
-                    .NotBeNull("because closing the passive side creates an error on the active one");
-            }
-        }
-
-        [Fact]
-        public void TestP2PChangeListener()
-        {
-            var statuses = new List<ReplicatorActivityLevel>();
-            var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, ProtocolType.ByteStream));
-            var awaiter = new ListenerAwaiter(listener);
-            var serverConnection = new MockServerConnection(listener, ProtocolType.ByteStream);
-            var config = new ReplicatorConfiguration(Db,
-                new MessageEndpoint("p2ptest1", serverConnection, ProtocolType.ByteStream,
-                    new MockConnectionFactory(null))) {
-                Continuous = true
-            };
-            listener.AddChangeListener((sender, args) => {
-                statuses.Add(args.Status.Activity);
-            });
-            var connection = listener.Connections;
-            RunReplication(config, 0, 0);
-            awaiter.WaitHandle.WaitOne(TimeSpan.FromSeconds(10)).Should().BeTrue();
-            awaiter.Validate();
-            statuses.Count.Should()
-                .BeGreaterThan(1, "because otherwise there were no callbacks to the change listener");
-        }
-
-        [Fact]
-        public void TestRemoveChangeListener()
-        {
-            var statuses = new List<ReplicatorActivityLevel>();
-            var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, ProtocolType.ByteStream));
-            var awaiter = new ListenerAwaiter(listener);
-            var serverConnection = new MockServerConnection(listener, ProtocolType.ByteStream);
-            var config = new ReplicatorConfiguration(Db,
-                new MessageEndpoint("p2ptest1", serverConnection, ProtocolType.ByteStream,
-                    new MockConnectionFactory(null))) {
-                Continuous = true
-            };
-            var token = listener.AddChangeListener((sender, args) => {
-                statuses.Add(args.Status.Activity);
-            });
-            var connection = listener.Connections;
-            listener.RemoveChangeListener(token);
-            RunReplication(config, 0, 0);
-            awaiter.WaitHandle.WaitOne(TimeSpan.FromSeconds(10)).Should().BeTrue();
-            awaiter.Validate();
-
-            statuses.Count.Should().Be(0);
-        }
-        
-        #endregion
 
         #region conflict resolving tests
 
@@ -1966,202 +1673,6 @@ namespace Test
 
         #endregion
 
-        #region URLEndpointListener tests
-
-#if COUCHBASE_ENTERPRISE
-
-        [Fact]
-        public void TestPort()
-        {
-            int exCnt = 0;
-
-            _config = new URLEndpointListenerConfiguration(_otherDB);
-            _config.Port = WSPort;
-            _config.DisableTLS = true;
-
-            //init a listener
-            _listener = new URLEndpointListener(_config);
-            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener has not yet started.");
-
-            try {
-                //start the listener
-                _listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                _listener.Port.Should().NotBe(0);
-                //stop the listener
-                _listener.Stop();
-                _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener is stopped.");
-            }
-        }
-
-        [Fact]
-        public void TestEmptyPort()
-        {
-            int exCnt = 0;
-
-            _config = new URLEndpointListenerConfiguration(_otherDB);
-            _config.Port = 0;
-            _config.DisableTLS = true;
-
-            //init a listener
-            _listener = new URLEndpointListener(_config);
-            _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener has not yet started.");
-
-            try {
-                //start the listener
-                _listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                _listener.Port.Should().NotBe(0);
-                //stop the listener
-                _listener.Stop();
-                _listener.Port.Should().Be(0, "Listener's port should be 0 because the listener is stopped.");
-            }
-        }
-
-        [Fact]
-        public void TestBusyPort()
-        {
-            CouchbasePosixException expectedException = null;
-            var listener = ListenerWithTLS(false, null);
-
-            var config = new URLEndpointListenerConfiguration(_otherDB);
-            config.Port = listener.Port;
-            config.DisableTLS = true;
-            var listener1 = new URLEndpointListener(config);
-
-            try {
-                listener1.Start();
-            } catch (CouchbasePosixException ex) {
-                expectedException = ex;
-            } finally {
-                listener.Stop();
-                listener1.Stop();
-            }
-
-            expectedException.Domain.Should().Be(CouchbaseLiteErrorType.POSIX);
-            expectedException.Error.Should().Be(PosixBase.GetCode(nameof(PosixWindows.EADDRINUSE)));
-        }
-
-        [Fact]
-        public void TestUrls()
-        {
-            int exCnt = 0;
-            var config = new URLEndpointListenerConfiguration(_otherDB);
-            var listener = new URLEndpointListener(config);
-            listener.Urls.Count.Should().Be(0);
-
-            try {
-                listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                listener.Urls.Count.Should().NotBe(0);
-                listener.Stop();
-                listener.Urls.Count.Should().Be(0);
-            }
-        }
-
-        [Fact]
-        public void TestStatus()
-        {
-            int exCnt = 0;
-            HashSet<ulong> maxConnectionCount = new HashSet<ulong>(),
-                maxActiveCount = new HashSet<ulong>();
-
-            _config = new URLEndpointListenerConfiguration(_otherDB);
-            _config.Port = WSPort;
-            _config.DisableTLS = true;
-            _config.Authenticator = null;
-
-            //init a listener
-            _listener = new URLEndpointListener(_config);
-            _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because the listener has not yet started.");
-            _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because the listener has not yet started.");
-
-            try {
-                //start the listener
-                _listener.Start();
-            } catch {
-                exCnt++;
-            } finally {
-                exCnt.Should().Be(0, "Because listener start should work without exception thrown.");
-                _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because no client connection has been established.");
-                _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because no client connection has been established.");
-
-                using (var doc1 = new MutableDocument("doc1"))
-                using (var doc2 = new MutableDocument("doc2")) {
-                    doc1.SetString("name", "Sam");
-                    Db.Save(doc1);
-                    doc2.SetString("name", "Mary");
-                    _otherDB.Save(doc2);
-                }
-
-                var targetEndpoint = new URLEndpoint(new Uri($"{_listener.Urls[0]}".Replace("http", "ws")));
-                var config = new ReplicatorConfiguration(Db, targetEndpoint);
-                using (var repl = new Replicator(config)) {
-                    var waitAssert = new WaitAssert();
-                    var token = repl.AddChangeListener((sender, args) =>
-                    {
-                        waitAssert.RunConditionalAssert(() =>
-                        {
-                            maxConnectionCount.Add(_listener.Status.ConnectionCount);
-                            maxActiveCount.Add(_listener.Status.ActiveConnectionCount);
-
-                            return args.Status.Activity == ReplicatorActivityLevel.Stopped;
-                        });
-                    });
-
-                    repl.Start();
-                    try {
-                        waitAssert.WaitForResult(TimeSpan.FromSeconds(100));
-                    } finally {
-                        repl.RemoveChangeListener(token);
-                    }
-                }
-
-                maxConnectionCount.Max().Should().Be(1);
-                maxActiveCount.Max().Should().Be(1);
-
-                //stop the listener
-                _listener.Stop();
-                _listener.Status.ConnectionCount.Should().Be(0, "Listener's connection count should be 0 because the connection is stopped.");
-                _listener.Status.ActiveConnectionCount.Should().Be(0, "Listener's active connection count should be 0 because the connection is stopped.");
-            }
-        }
-
-        private URLEndpointListener ListenerWithTLS(bool tls, IListenerAuthenticator auth)
-        {
-            int exCnt = 0;
-            var config = new URLEndpointListenerConfiguration(Db);
-            config.Port = tls ? WSSPort : WSPort;
-            config.DisableTLS = !tls;
-            config.Authenticator = auth;
-
-            var listener = new URLEndpointListener(config);
-
-            try {
-                //start the listener
-                listener.Start();
-            } catch {
-                exCnt++;
-            }
-
-            if (exCnt == 0)
-                return listener;
-            else
-                return null;
-        }
-
-#endif
-        #endregion
-
         private void WithActiveReplicationAndQuery(bool isCloseNotDelete)
         {
             Database.Delete("closeDB", Db.Config.Directory);
@@ -2415,132 +1926,6 @@ namespace Test
             }
         }
 
-        private ReplicatorConfiguration CreateFailureP2PConfiguration(ProtocolType protocolType, MockConnectionLifecycleLocation location, bool recoverable)
-        {
-            var errorLocation = TestErrorLogic.FailWhen(location);
-
-            if (recoverable) {
-                errorLocation.WithRecoverableException();
-            } else {
-                errorLocation.WithPermanentException();
-            }
-
-            var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, protocolType));
-            var server = new MockServerConnection(listener, protocolType) {
-                ErrorLogic = errorLocation
-            };
-
-            var config = new ReplicatorConfiguration(Db,
-                new MessageEndpoint("p2ptest1", server, protocolType, new MockConnectionFactory(errorLocation))) {
-                ReplicatorType = ReplicatorType.Push,
-                Continuous = false
-            };
-
-            var connection = listener.Connections;
-            return config;
-        }
-
-        private void TestP2PError(MockConnectionLifecycleLocation location, bool recoverable)
-        {
-            using (var mdoc = new MutableDocument("livesindb")) {
-                mdoc.SetString("name", "db");
-                Db.Save(mdoc);
-            }
-
-            var expectedDomain = recoverable ? 0 : CouchbaseLiteErrorType.CouchbaseLite;
-            var expectedCode = recoverable ? 0 : (int)CouchbaseLiteError.WebSocketUserPermanent;
-
-            var config = CreateFailureP2PConfiguration(ProtocolType.ByteStream, location, recoverable);
-            RunReplication(config, expectedCode, expectedDomain);
-            config = CreateFailureP2PConfiguration(ProtocolType.MessageStream, location, recoverable);
-            RunReplication(config, expectedCode, expectedDomain, true);
-        }
-
-        private void RunTwoStepContinuous(ReplicatorType type, string uid)
-        {
-            var listener = new MessageEndpointListener(new MessageEndpointListenerConfiguration(_otherDB, ProtocolType.ByteStream));
-            var server = new MockServerConnection(listener, ProtocolType.ByteStream);
-            var config = new ReplicatorConfiguration(Db,
-                new MessageEndpoint(uid, server, ProtocolType.ByteStream, new MockConnectionFactory(null))) {
-                ReplicatorType = type,
-                Continuous = true
-            };
-
-            using (var replicator = new Replicator(config)) {
-                replicator.Start();
-
-                Database firstSource = null;
-                Database secondSource = null;
-                Database firstTarget = null;
-                Database secondTarget = null;
-                if (type == ReplicatorType.Push) {
-                    firstSource = Db;
-                    secondSource = Db;
-                    firstTarget = _otherDB;
-                    secondTarget = _otherDB;
-                } else if (type == ReplicatorType.Pull) {
-                    firstSource = _otherDB;
-                    secondSource = _otherDB;
-                    firstTarget = Db;
-                    secondTarget = Db;
-                } else {
-                    firstSource = Db;
-                    secondSource = _otherDB;
-                    firstTarget = _otherDB;
-                    secondTarget = Db;
-                }
-
-                using (var mdoc = new MutableDocument("livesindb")) {
-                    mdoc.SetString("name", "db");
-                    mdoc.SetInt("version", 1);
-                    firstSource.Save(mdoc);
-                }
-
-                var count = 0;
-                while (replicator.Status.Progress.Completed == 0 ||
-                       replicator.Status.Activity != ReplicatorActivityLevel.Idle) {
-                    Thread.Sleep(500);
-                    count++;
-                    count.Should().BeLessThan(10, "because otherwise the replicator did not advance");
-                }
-
-                var previousCompleted = replicator.Status.Progress.Completed;
-                firstTarget.Count.Should().Be(1);
-
-                using (var savedDoc = secondSource.GetDocument("livesindb"))
-                using (var mdoc = savedDoc.ToMutable()) {
-                    mdoc.SetInt("version", 2);
-                    secondSource.Save(mdoc);
-                }
-
-                count = 0;
-                while (replicator.Status.Progress.Completed == previousCompleted ||
-                       replicator.Status.Activity != ReplicatorActivityLevel.Idle) {
-                    Thread.Sleep(500);
-                    count++;
-                    count.Should().BeLessThan(10, "because otherwise the replicator did not advance");
-                }
-
-                using (var savedDoc = secondTarget.GetDocument("livesindb")) {
-                    savedDoc.GetInt("version").Should().Be(2);
-                }
-
-                replicator.Stop();
-                while (replicator.Status.Activity != ReplicatorActivityLevel.Stopped) {
-                    Thread.Sleep(100);
-                }
-            }
-        }
-
-#endif
-
-#if COUCHBASE_ENTERPRISE
-        private ReplicatorConfiguration CreateConfig(bool push, bool pull, bool continuous)
-        {
-            var target = _otherDB;
-            return CreateConfig(push, pull, continuous, target);
-        }
-
         private void TestPushDocWithFilter(bool continuous)
         {
             using (var doc1 = new MutableDocument("doc1"))
@@ -2560,15 +1945,14 @@ namespace Test
             _otherDB.GetDocument("doc2").Should().NotBeNull("because doc2 is filtered in in the callback");
             _isFilteredCallback = false;
         }
-#endif
-
-        private ReplicatorConfiguration CreateConfig(bool push, bool pull, bool continuous, URLEndpoint endpoint)
-        {
-            var retVal = new ReplicatorConfiguration(Db, endpoint);
-            return ModifyConfig(retVal, push, pull, continuous);
-        }
 
 #if COUCHBASE_ENTERPRISE
+        private ReplicatorConfiguration CreateConfig(bool push, bool pull, bool continuous)
+        {
+            var target = _otherDB;
+            return CreateConfig(push, pull, continuous, target);
+        }
+
         private ReplicatorConfiguration CreateConfig(bool push, bool pull, bool continuous, Database target)
         {
             var retVal = new ReplicatorConfiguration(Db, new DatabaseEndpoint(target));
